Changed file structure; moving to multiple implementations of bit_builder.
