+++ .github/workflows/run_tests.yml
@@ -83,7 +83,7 @@ jobs:
             blueprint_algebra_fields_plonk_non_native_bit_decomposition_test,
             blueprint_algebra_fields_plonk_non_native_bit_composition_test,
             blueprint_algebra_fields_plonk_non_native_bit_shift_test,
-            blueprint_algebra_fields_plonk_non_native_boolean_logic_ops_test,
+            blueprint_algebra_fields_plonk_non_native_logic_ops_test,
             blueprint_non_native_plonk_scalar_non_native_range_test,
             blueprint_non_native_plonk_bool_scalar_multiplication_test,
             blueprint_hashes_plonk_decomposition_test,
