+++ test/CMakeLists.txt
@@ -59,7 +59,7 @@ set(NON_NATIVE_TESTS_FILES
     "algebra/fields/plonk/non_native/bit_decomposition"
     "algebra/fields/plonk/non_native/bit_composition"
     "algebra/fields/plonk/non_native/bit_shift"
-    "algebra/fields/plonk/non_native/boolean/logic_ops"
+    "algebra/fields/plonk/non_native/logic_ops"
     "non_native/plonk/bool_scalar_multiplication"
     "non_native/plonk/scalar_non_native_range"
     )
